Convert KRandom seed to long int.

+++ Source/Kassiopeia/LMCRunPause.hh
@@ -62,7 +62,7 @@ namespace locust
             bool AddMaxTimeTerminator( const scarab::param_node& aParam );
             bool AddMaxRTerminator( const scarab::param_node& aParam );
             bool AddGenerator( const scarab::param_node& aParam );
-            int GetSeed( const scarab::param_node& aParam );
+            long int GetSeed( const scarab::param_node& aParam );
